moved mount check to log_folder.sh, updated installer

+++ helpers/log_folder.sh
@@ -24,8 +24,24 @@ log_folder()
     read -p "(y/n)  " -r ANSWER
     case ${ANSWER:0:1} in
         y|Y|"" )
-          rm -R "${LOG_DIR:?}/"*
-          echo -e "\\n${GREEN}Sucessfully deleted: $LOG_DIR ${NC}\\n"
+          if mount | grep "$LOG_DIR" | grep -e "proc\|sys\|run" > /dev/null; then
+            echo
+            print_output "[!] We found unmounted areas from a former emulation process in your log directory $LOG_DIR." "no_log"
+            print_output "[!] You should unmount this stuff manually:\\n" "no_log"
+            mount | grep "$LOG_DIR"
+            echo -e "\\n${RED}Terminate emba${NC}\\n"
+            exit 1
+          elif mount | grep "$LOG_DIR" > /dev/null; then
+            echo
+            print_output "[!] We found unmounted areas in your log directory $LOG_DIR." "no_log"
+            print_output "[!] You should unmount this stuff manually:\\n" "no_log"
+            mount | grep "$LOG_DIR"
+            echo -e "\\n${RED}Terminate emba${NC}\\n"
+            exit 1
+          else
+            rm -R "${LOG_DIR:?}/"*
+            echo -e "\\n${GREEN}Sucessfully deleted: $LOG_DIR ${NC}\\n"
+          fi
         ;;
         * )
           echo -e "\\n${RED}Terminate emba${NC}\\n"
